requirements.txt
xentocanto-api  
pandas
numpy
urllib3<2
librosa
matplotlib
torch 
torchvision 
pandas 
scikit-learn 
pillow
seaborn